cuda: fixes problems with NS(CudaContext*) APIs
